# Fix ADC Missed 1 Bit Error