# Description of Changes
I found this when I was trying to add a register to read out the ADC data for artiq_sinara_tester for production test. There is one bit that is missed and not read for each transaction. This is due to SCK not being idle HIGH as specified in the LTC2323-16 datasheet. See below for reference.

![image](https://github.com/user-attachments/assets/4d6faa64-bb95-4e5b-9b7a-5821df87f8d0)
Note: Timing Diagram from LTC2323-16 datasheet

Since the ADC data is clocked out at the falling edge, at the end of transfer cycle, there is a bit missed and not read by the gateware. See the captured example waveform below.

ADC should read around -557.875uV ->0xEE91. If you also consider the last bit that is not read, that would be 0xEE91.
![image](https://github.com/user-attachments/assets/3b037049-89d0-4ce3-a993-5f642c04c6c2)
Note: SCK is slowed down to account for the logic analyzer bandwidth.

This PR fixes this issue by flipping the output of SCK. This changes also work in its original SCK speed. Below shows the captured waveform after the fix.
![image](https://github.com/user-attachments/assets/e3b97b35-0ab6-40d9-a2de-27a5e606ca0f)
Note: SCK is slowed down to account for the logic analyzer bandwidth.