@linuswck 2025-06-13
I typed the unit wrongly. Should be mV instead of uV. 
The original code has SCK idle low, which is not the same as specified in the timing diagram. 

I have edited the PR message with clearer descriptions.