@jordens 2025-06-13
Thanks for the revised analysis.

I'm not happy with the timing on that tDCNVSCKL. We want to ensure it's at least 10 ns plus the prop delay of the CNV driver by looking at the code! Since the SCK cycle is 10 ns in the code, there needs to be at least another 4 ns delay cycle between the CNV assertion and the start of the SCK shifting. Either through the FSM logic or buffer/IO registers.

Also you don't make an attempt to argue why your change is the correct fix. You need to do that, otherwise it's accidental. Why not change the synthesizer reset value? Is that value correct?

Don't look at your scope so much. The sample rate is set in the code. It was either wrong before and/or it's wrong now. Can you check the t_read calculation and whether it meets the timing requirements now?

Regarding the sampling edge: "proof by example" doesn't work. Correct values are inconclusive. You have to analyze the code and the timings. If we touch this code, I would like the design to be correct.