@linuswck 2025-06-16
**Correction**: SCK should be 250MHz / 6 -> 83.3MHz. (Double Checked on oscilloscope) 
The original comment in adc.py is misleading. I have corrected that.

The original code was designed to be idle high on simulation. I think the original author forgot that the polarity of `SCK` signal is **swapped** at the [DiffrentialOutput](https://github.com/quartiq/phaser/blob/72ec22a10d8972ffd9bf22f4f9504b6bf3c2ed88/adc.py#L57) primitives. Setting the `ddr_clk_synth` reset value to `0b111000` has the same effect. Flipping the polarity of `SCK` should be the correct change to be made.

For the measurement I took, I used two Rigol Pvp2350 Probe, which has an 50 pF rated capacitance, to probe the IC24's Output. Looking at the datasheet, the fall time is way out of spec. Therefore, without probing the IC24's Output. the fall time should be way better.
![image](https://github.com/user-attachments/assets/53625533-6844-484c-a707-258b256a0b28)

But, I agree with you that I should take a closer look at the timing. Previously, I didn't considered the IOB Delay. Turns out in the boundary cases, `CNV_N` relative to the `SCK` in worse case scenario can be arrived at the IC slower by ~4 ns.

![image](https://github.com/user-attachments/assets/103ee75d-993c-4c3a-ad46-0755920bb419)

If I also take the IOB delay into account, t_DCNVSCKL will be violated if I do not adjust the `t_conv` & `t_rtt` value.
I change the following values should meet the timings in all scenarios.
t_conv: 3 -> 4 & t_rtt: 6 -> 5.

# Timings 
`t_read` counter value is responsible for generating the required number of `SCK` cycles to clock out the register data. Since the PR only flips the `SCK` polarity, there is no timing changes related to `t_read` during the transaction. `adc.done` is issued just a sys clk cycle before the deassertion of `CNVN`. Given the long `t_rtt` time, the `SDO` should have been sampled correctly already before it is processed by other gateware logic. 

Looking at the datasheet, swapping the polarity of `SCK` should only affect `t_DCNVSCKL`, `tDCNVSDOZ` timings.

## Timing Requirements
SN65LVDS2DBVR propagation delay @ 30 degree = `CNVN_Delay` = 2.65 ns 
t_DCNVSCKL = 1st SCK Falling Edge from CNVN assertion >= 9.5 ns
t_DCNVSDOZ = Last SCK Falling Edge to CNVN deassertion >= 19.1 ns

t_DCNVSDOZ should have no problem after flipping the SCK polarity. So, let's consider the worse case scenario for t_DCNVSCKL, let's name the delay to be `ODDR_Delay` = 3.836+0.113 ~ 4 ns  

## Expected Timings Before Change
t_DCNVSCKL = t_conv * sys_period + SCK_period/2 + `ODDR_Delay` - `CNVN_Delay` = 3 * 4 + 6 - `ODDR_Delay` - `CNVN_Delay` = 11.35 ns
t_DCNVSDOZ = t_rtt * sys_period +  SCK_period/2 + `ODDR_Delay` + `CNVN_Delay`= 6 * 4 + 6 + `ODDR_Delay` + `CNVN_Delay` = 36.65 ns

## Expected Timings After Change
t_DCNVSCKL = t_conv * sys_period + sys_period + `ODDR_Delay` - `CNVN_Delay`= 4 * 4 + 4 - `ODDR_Delay` - `CNVN_Delay` = 13.35 ns.
t_DCNVSDOZ  = t_rtt * sys_period + SCK_Period + `ODDR_Delay` + `CNVN_Delay`= 5 * 4 + 12 + `ODDR_Delay` + `CNVN_Delay` = 38.65 ns

# Simulations
I changed `t_conv` =4 & `t_rtt` = 5 for the upcoming simulations. Other parameters remain unchanged. For the sck to be visible on the timing simulation, I connect an ODDR like block with a **fixed** delay of  **1 ns**. For the actual timing calculations, we should take the `report_datasheet` values.

DO NOTE that as we cannot simulate any output primitives, the polarity `SCK` shown in the simulations should have been [**swapped**](https://github.com/quartiq/phaser/blob/72ec22a10d8972ffd9bf22f4f9504b6bf3c2ed88/adc.py#L57) when it got deployed on the hardware.

## Original Code
![image](https://github.com/user-attachments/assets/1a8178bb-f63f-414c-86b7-c0d7fe326c88)
t_DCNVSCKL  = 64 ns - 42 ns - 4 ns - 2.65 ns = 15.35 ns
t_DCNVSDOZ = 274 ns - 244 ns + 4 ns + 2.65 ns = 36.65 ns

## PR Code
![image](https://github.com/user-attachments/assets/4a3fa143-3b4c-4fd3-98a0-86b3b452100b)
t_DCNVSCKL  = 62 ns - 42 ns - 4 ns - 2.65 ns = 13.35 ns
t_DCNVSDOZ = 274 ns - 242 ns + 4 ns + 2.65 ns = 38.65 ns 

## Set the ddr_clk_synth Reset Value to `0b111000`
It shows same CNVN and SCK waveform as the above PR Code.