@jordens 2025-06-30
> Correction: SCK should be 250MHz / 6 -> 83.3MHz. (Double Checked on oscilloscope)

Certainly not.

> Setting the ddr_clk_synth reset value to 0b111000 has the same effect. 

Do that then! Less logic, fewer inversions, simpler.

Changing t_rtt needs a rationale as well. t_rtt is not just for tDSCKLCNVH but also for synchronizing the ret->sys crossing (not for sdo sampling, that's done by the source synchronous design). Changing clock polarity then has an affect on both factors.